Specified exception - throw situations

+++ CoffeeMaker/src/exceptions/WaterException.java
@@ -4,7 +4,7 @@ public class WaterException extends Exception {
 
     private int waterCups;
 
-    public WaterException(int waterCups){
+    public WaterException(double waterCups){
         super(String.format("%s too much water!!", waterCups));
     }
 }
